+++ src/Menu.php
@@ -22,11 +22,11 @@ use PhpSchool\CliMenu\Builder\CliMenuBuilder;
 class Menu extends CliMenuBuilder
 {
     /**
-     * The current option selected, if any.
+     * The current option value.
      *
-     * @var \NunoMaduro\LaravelConsoleMenu\SelectedOption
+     * @var mixed
      */
-    private $optionSelected;
+    private $result;
 
     /**
      * Menu constructor.
@@ -61,7 +61,7 @@ class Menu extends CliMenuBuilder
                 $value,
                 $label,
                 function (CliMenu $menu) {
-                    $this->optionSelected = $menu->getSelectedItem();
+                    $this->result = $menu->getSelectedItem()->fetch();
                     $menu->close();
                 }
             )
@@ -87,22 +87,21 @@ class Menu extends CliMenuBuilder
     }
 
     /**
-     * Adds a new text input option.
+     * Add a question.
      *
      * @param string $label
-     * @param string $promptText
      * @param string $placeholder
      * @return Menu
      */
-    public function addTextInputOption(string $label, string $promptText, string $placeholder = 'Enter text...'): Menu
+    public function addQuestion(string $label, string $placeholder = ''): Menu
     {
-        $itemCallable = function (CliMenu $menu) use ($promptText, $placeholder) {
+        $itemCallable = function (CliMenu $menu) use ($label, $placeholder) {
             $result = $menu->askText()
-                ->setPromptText($promptText)
+                ->setPromptText($label)
                 ->setPlaceholderText($placeholder)
                 ->ask();
 
-            $this->optionSelected = new InputTextOption($result->fetch());
+            $this->result = $result->fetch();
 
             $menu->close();
         };
@@ -113,7 +112,7 @@ class Menu extends CliMenuBuilder
     }
 
     /**
-     * Opens the menu and returns the selected option value.
+     * Open the menu and return the result.
      *
      * @return mixed
      */
@@ -122,6 +121,19 @@ class Menu extends CliMenuBuilder
         $this->build()
             ->open();
 
-        return $this->optionSelected ? $this->optionSelected->getValue() : null;
+        return $this->result;
+    }
+
+    /**
+     * Set the result.
+     *
+     * @param mixed $result
+     * @return Menu
+     */
+    public function setResult($result): Menu
+    {
+        $this->result = $result;
+
+        return $this;
     }
 }