+++ CHANGELOG.md
@@ -6,11 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.0.0] - 2018-07-30
 ### Changed
-- Using `"php-school/cli-menu": "^3.0"`
-- Update namespace on CliMenuBuilder
+- Upgraded Cli Menu to v3.
 
 ### Added
-- Added Menu#addTextInputOption
+- `addQuestion` method to `Menu`.
+- `setResult` method to `Menu`, to allow using extra options.
    
 ## [1.0.3] - 2018-05-08
 ### Changed